Move `getTemplateVersionByHash` out of shared util
